cleanup &  buildable on linux, rename a few things
