+++ DarkLogiXBrowser.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31912.275
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DarkLogiXBrowser", "TestPatch\DarkLogiXBrowser.csproj", "{03E7BB86-FE1F-4A2D-A176-B4A6AC01830A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DarkLogiXBrowser", "DarkLogiXBrowser\DarkLogiXBrowser.csproj", "{03E7BB86-FE1F-4A2D-A176-B4A6AC01830A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
