--- TestPatch/DarkLogiXBrowser.csproj.user
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'">
-    <StartAction>Program</StartAction>
-    <StartProgram>C:\Neos\app\Neos.exe</StartProgram>
-    <StartArguments>-DontAutoOpenCloudHome -SkipIntroTutorial -Screen -LoadAssembly "C:\Neos\app\Libraries\NeosModLoader.dll"</StartArguments>
-    <StartWorkingDirectory>C:\Neos\app\</StartWorkingDirectory>
-  </PropertyGroup>
-</Project>
